// src/app/models/churrascoNovo.ts
export class ChurrascoNovo {
   descricao: string;
   carneBovina: boolean;
   id_usuario: string;
   carneSuina: boolean;
   carneFrango: boolean;
   saladaVerde: boolean;
   arroz: boolean;
   maionese: boolean;
   pao: boolean;
   quantidade_pessoas: number;
   quantidade_carne: number;
   valor_total: number;
   data: Date;
}